Moved Batch backend configs
